refactoring code with modules
