+++ modules/error.js
@@ -1,0 +1,5 @@
+// create an error mssg
+const error = document.createElement('span');
+error.innerHTML = 'please fill out all these fields';
+
+export {error};
